fix(table/web) : Added makeStories and modified Table story
